Add very basics for a flat model
This model won't allow anything but direct values. In context of a tree,
this model would always be a collection of leaf values. For now, all
values are always boxed, meaning they're not stored efficiently per se.
This may be revised at a later point, because right now, because of the
boxing, it's not possible either to extract a `short` as a `int` if you
wanted to.