Use regex to validate keys in flat model
Instead of allowing any but whitespace characters is maybe a bit too
loose for now. Therefore restricting it to alphanumerical only at this
moment. Because of that, only one type of `Exception` can be thrown, so
the tests had to be adapted.

Sadly, `ArgumentException` will automatically append the parameter name
to its `Message` property. That's why we use `StartsWith` instead.
Additionally, `TUnit` seems to have introduced a bug with `0.22.31`,
where `ThrowsAsync` doesn't check the type on itself any more. That's
why it was reverted back to `0.22.24`, the last version before it.